[UiBundle] Move classes from Command to Console/Command
