+++ src/Sylius/Bundle/UiBundle/Command/DebugTemplateEventCommand.php
@@ -13,59 +13,18 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\UiBundle\Command;
 
-use Sylius\Bundle\UiBundle\Registry\TemplateBlock;
-use Sylius\Bundle\UiBundle\Registry\TemplateBlockRegistryInterface;
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Style\SymfonyStyle;
+trigger_deprecation(
+    'sylius/ui-bundle',
+    '1.13',
+    'The "%s" class is deprecated and will be removed in Sylius 2.0. Use "%s" instead.',
+    DebugTemplateEventCommand::class,
+    \Sylius\Bundle\UiBundle\Console\Command\DebugTemplateEventCommand::class,
+);
 
-final class DebugTemplateEventCommand extends Command
-{
-    protected static $defaultName = 'sylius:debug:template-event';
+class_exists(\Sylius\Bundle\UiBundle\Console\Command\DebugTemplateEventCommand::class);
 
-    public function __construct(private TemplateBlockRegistryInterface $templateBlockRegistry)
+if (false) {
+    final class DebugTemplateEventCommand
     {
-        parent::__construct();
-    }
-
-    protected function configure(): void
-    {
-        $this
-            ->setDescription('Debug template events and associated blocks')
-            ->addArgument('event', InputArgument::OPTIONAL, 'Template event name', null)
-        ;
-    }
-
-    protected function execute(InputInterface $input, OutputInterface $output): int
-    {
-        $io = new SymfonyStyle($input, $output);
-        $eventName = $input->getArgument('event');
-
-        if ($eventName === null) {
-            $io->title('List of template events');
-
-            $io->listing(array_keys($this->templateBlockRegistry->all()));
-
-            return 0;
-        }
-
-        $io->title(sprintf('Blocks registered for the template event "%s"', $eventName));
-
-        $io->table(
-            ['Block name', 'Template', 'Priority', 'Enabled'],
-            array_map(
-                static fn (TemplateBlock $templateBlock): array => [
-                    $templateBlock->getName(),
-                    $templateBlock->getTemplate(),
-                    $templateBlock->getPriority(),
-                    $templateBlock->isEnabled() ? 'TRUE' : 'FALSE',
-                ],
-                $this->templateBlockRegistry->all()[$eventName] ?? [],
-            ),
-        );
-
-        return 0;
     }
 }
