+++ src/Sylius/Bundle/UiBundle/Resources/config/services/template_event.xml
@@ -27,9 +27,11 @@
             <argument type="service" id="Sylius\Bundle\UiBundle\Renderer\TemplateBlockRendererInterface" />
         </service>
 
-        <service id="Sylius\Bundle\UiBundle\Command\DebugTemplateEventCommand">
+        <service id="Sylius\Bundle\UiBundle\Console\Command\DebugTemplateEventCommand">
             <argument type="service" id="Sylius\Bundle\UiBundle\Registry\TemplateBlockRegistryInterface" />
             <tag name="console.command" command="sylius:debug:template-event" />
         </service>
+
+        <service id="Sylius\Bundle\UiBundle\Command\DebugTemplateEventCommand" alias="Sylius\Bundle\UiBundle\Console\Command\DebugTemplateEventCommand" />
     </services>
 </container>
